requirements.txt
# Core app
streamlit
streamlit-float
PyMuPDF
python-dotenv
requests
pandas
numpy

# Google Gemini API
google-generativeai

# LangChain ecosystem
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.79
openai
tiktoken
faiss-cpu
sentence-transformers
rank-bm25
fpdf
# ML + Transformers
torch
torchvision
torchaudio
transformers
scikit-learn

# Utilities
tenacity
python-dateutil
PyYAML
tiktoken
streamlit-chat-widget
./Chat_input_widget-master